Проект разобран на модули
