+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/gitHub/IGitHubService.kt
@@ -1,0 +1,10 @@
+package ru.appdevelopers.domain.gitHub
+
+import io.reactivex.Single
+import ru.appdevelopers.domain.gitHub.model.GitHubAccessToken
+import ru.appdevelopers.domain.models.ApiResponse
+
+
+interface IGitHubService {
+    fun auth(code: String): Single<ApiResponse<GitHubAccessToken>>
+}
