+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/gitHub/model/GitHubAccessToken.kt
@@ -1,0 +1,8 @@
+package ru.appdevelopers.domain.gitHub.model
+
+import ru.appdevelopers.domain.models.AccessToken
+
+class GitHubAccessToken(
+    accessToken: String? = null,
+    username: String? = null
+): AccessToken(accessToken, username)
