+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/gitHub/model/GitHubAccessTokenResponse.kt
@@ -1,0 +1,11 @@
+package ru.appdevelopers.domain.gitHub.model
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * Ответ с токеном авторизации через гитхаб
+ */
+data class GitHubAccessTokenResponse(
+    @SerializedName("access_token")
+    val accessToken: String
+)
