+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/googleAuth/model/GoogleAccessToken.kt
@@ -1,0 +1,8 @@
+package ru.appdevelopers.domain.googleAuth.model
+
+import ru.appdevelopers.domain.models.AccessToken
+
+class GoogleAccessToken(
+    accessToken: String? = null,
+    username: String? = null
+): AccessToken(accessToken, username)
