+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/models/AccessToken.kt
@@ -1,0 +1,6 @@
+package ru.appdevelopers.domain.models
+
+open class AccessToken(
+    open val accessToken: String? = null,
+    open val username: String? = null
+)
