+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/models/ErrorEvent.kt
@@ -1,0 +1,5 @@
+package ru.appdevelopers.domain.models
+
+sealed class AuthErrorResponse
+class GoogleAuthErrorResponse<T>(val data: T) : AuthErrorResponse()
+class GitHubAuthErrorResponse<T>(val data: T) : AuthErrorResponse()
