+++ GitHubClient/app/domain/src/main/java/ru/appdevelopers/domain/models/User.kt
@@ -1,0 +1,3 @@
+package ru.appdevelopers.domain.models
+
+data class User(val type: AuthType, val userName: String? = null)
