+++ GitHubClient/app/googleAuth/build.gradle
@@ -1,7 +1,6 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'kotlin-android'
-    id 'kotlin-android-extensions'
     id 'kotlin-kapt'
 }
 
@@ -9,21 +8,16 @@ android {
     compileSdkVersion 30
 
     defaultConfig {
-        applicationId "ru.appdevelopers.githubclient"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
     }
 
     buildTypes {
-        debug {
-
-        }
-
-
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
@@ -41,31 +35,17 @@ android {
 dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.5.0'
+    implementation 'androidx.core:core-ktx:1.6.0'
     implementation 'androidx.appcompat:appcompat:1.3.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'com.google.android.gms:play-services-auth:19.0.0'
+    implementation 'com.google.android.material:material:1.4.0'
     testImplementation 'junit:junit:4.+'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-
-    implementation "io.reactivex.rxjava2:rxandroid:2.0.1"
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation 'com.google.android.gms:play-services-auth:19.0.0'
 
     // di
     implementation 'com.github.stephanenicolas.toothpick:ktp:3.0.0'
     kapt 'com.github.stephanenicolas.toothpick:toothpick-compiler:3.0.0'
 
-    // Moxy
-    implementation "com.github.moxy-community:moxy:2.2.0"
-    implementation "com.github.moxy-community:moxy-androidx:2.2.0"
-    kapt "com.github.moxy-community:moxy-compiler:2.2.0"
-
-    // Cicerone (Navigation pattern)
-    implementation("com.github.terrakok:cicerone:7.0")
-
-    implementation project(':app:dal')
     implementation project(':app:domain')
-    implementation project(':app:gitHubService')
-    implementation project(':app:googleAuth')
 }
