--- GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/gitHub/model/GitHubAccessTokenResponse.kt
@@ -1,11 +1,0 @@
-package ru.appdevelopers.githubclient.domain.gitHub.model
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Ответ с токеном авторизации через гитхаб
- */
-data class GitHubAccessTokenResponse(
-    @SerializedName("access_token")
-    val accessToken: String
-)
