--- GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/googleAuth/model/GoogleAccessToken.kt
@@ -1,8 +1,0 @@
-package ru.appdevelopers.githubclient.domain.googleAuth.model
-
-import ru.appdevelopers.githubclient.domain.models.AccessToken
-
-class GoogleAccessToken(
-    accessToken: String? = null,
-    username: String? = null
-): AccessToken(accessToken, username)
