--- GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/models/AccessToken.kt
@@ -1,6 +1,0 @@
-package ru.appdevelopers.githubclient.domain.models
-
-open class AccessToken(
-    open val accessToken: String? = null,
-    open val username: String? = null
-)
