--- GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/models/ErrorEvent.kt
@@ -1,5 +1,0 @@
-package ru.appdevelopers.githubclient.domain.models
-
-sealed class AuthErrorResponse
-class GoogleAuthErrorResponse<T>(val data: T) : AuthErrorResponse()
-class GitHubAuthErrorResponse<T>(val data: T) : AuthErrorResponse()
