+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/ui/auth/AuthPresenter.kt
@@ -10,12 +10,12 @@ import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.observers.DisposableSingleObserver
 import io.reactivex.schedulers.Schedulers
 import moxy.InjectViewState
-import ru.appdevelopers.githubclient.domain.models.*
-import ru.appdevelopers.githubclient.data.repository.IUserRepository
-import ru.appdevelopers.githubclient.domain.gitHub.IGitHubService
-import ru.appdevelopers.githubclient.domain.gitHub.model.GitHubAccessToken
-import ru.appdevelopers.githubclient.domain.googleAuth.IAccessTokenMapper
-import ru.appdevelopers.githubclient.domain.googleAuth.model.GoogleAccessToken
+import ru.appdevelopers.dal.repository.IUserRepository
+import ru.appdevelopers.domain.models.*
+import ru.appdevelopers.domain.gitHub.IGitHubService
+import ru.appdevelopers.domain.gitHub.model.GitHubAccessToken
+import ru.appdevelopers.domain.googleAuth.IAccessTokenMapper
+import ru.appdevelopers.domain.googleAuth.model.GoogleAccessToken
 import ru.appdevelopers.githubclient.ui.Screens
 import ru.appdevelopers.githubclient.ui.base.BasePresenter
 import javax.inject.Inject
