+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/ui/auth/IAuthCallback.kt
@@ -1,8 +1,8 @@
 package ru.appdevelopers.githubclient.ui.auth
 
 import moxy.viewstate.strategy.alias.AddToEndSingle
-import ru.appdevelopers.githubclient.domain.models.AccessToken
-import ru.appdevelopers.githubclient.domain.models.AuthErrorResponse
+import ru.appdevelopers.domain.models.AccessToken
+import ru.appdevelopers.domain.models.AuthErrorResponse
 import ru.appdevelopers.githubclient.ui.IBaseCallback
 
 interface IAuthCallback : IBaseCallback {
